fixed a race condition in cheaper + threads

+++ master.c
@@ -1370,10 +1370,13 @@ int master_loop(char **argv, char **environ) {
 
 #endif
 						}
-						kill(uwsgi.workers[i].pid, SIGUSR2);
-						// allow SIGUSR2 to be delivered
-						sleep(1);
-						kill(uwsgi.workers[i].pid, SIGKILL);
+
+						if (uwsgi.workers[i].pid > 0) {
+							kill(uwsgi.workers[i].pid, SIGUSR2);
+							// allow SIGUSR2 to be delivered
+							sleep(1);
+							kill(uwsgi.workers[i].pid, SIGKILL);
+						}
 						// to avoid races
 						uwsgi.workers[i].harakiri = 0;
 					}
@@ -1674,6 +1677,7 @@ int master_loop(char **argv, char **environ) {
 			if (uwsgi.workers[uwsgi.mywid].cheaped == 1) {
 				uwsgi.workers[uwsgi.mywid].pid = 0;
 				uwsgi_log("uWSGI worker %d cheaped.\n", uwsgi.mywid);
+				uwsgi.workers[uwsgi.mywid].harakiri = 0;
 				continue;
 			}
 			gettimeofday(&last_respawn, NULL);
